Add album title to the applet.
It's not perfect, but it works.
There are two new options:
- showAlbumTitle
- albumBeneathSongAndArtists
one toggles the abum, the other moves it around.

Through a very terrifying amount of nested if / elses, the text organizes itself so that it never takes up three lines in the toolbar widget. However, this could be adjusted rather easily by editing said if / elses.

Added styles to the album name and song title that work regardless of the layout